+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### App Center Crashes
 
 * **[Feature]** Catch "low memory warning" and provide the API to check if it has happened in last session: `MSCrashes.hasReceivedMemoryWarningInLastSession()`.
+* **[Fix]** Fix main thread checker's warning during crash processing on macOS.
 
 ### App Center Distribute