+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # App Center SDK for iOS, macOS and tvOS Change Log
 
-## Version 2.2.1 (Under development)
+## Version 2.3.0 (Under development)
 
 ### App Center Crashes
 
+* **[Feature]** Catch "low memory warning" and provide the API to check if it has happened in last session: `MSCrashes.hasReceivedMemoryWarningInLastSession()`.
 * **[Fix]** Fix main thread checker's warning during crash processing on macOS.
 
 ### App Center Distribute